refactor: Separate Article component from App component
