+++ src/App.test.tsx
@@ -26,16 +26,6 @@ test(`should have ${LIST_LENGTH} li tags inside nav`, () => {
   expect(listItems.length).toBe(LIST_LENGTH);
 });
 
-test('Article tag should have "Hello, React Web" text and an h1 tag with text "Welcome"', () => {
-  render(<App />);
-
-  const headingElement = screen.getByRole('heading', { name: 'Welcome' });
-  expect(headingElement).toBeInTheDocument();
-
-  const helloText = screen.getByText('Hello, React Web');
-  expect(helloText).toBeInTheDocument();
-});
-
 test('Create Button test', () => {
   render(<App />);
