+++ src/components/article/Article.tsx
@@ -1,0 +1,11 @@
+import "./Article.scss";
+import { Topic } from "models/contents";
+
+export function Article(props: Omit<Topic, 'id'>): JSX.Element {
+    return (
+        <article>
+            <h1>{props.title}</h1>
+            {props.body}
+        </article>
+    );
+}
